Fixed a few signed vs unsigned warnings
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@39677 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/RandomGenerator.h
@@ -36,7 +36,7 @@ public:
 private:
 	const static size_t BUFFER_SIZE = 4096;
 
-	int bufferPos;
+	size_t bufferPos;
 	char buffer[BUFFER_SIZE];
 };
